refactor: split CLI around static component registries
The CLI had become the integration point for every managed lifecycle concern, which made omissions across install, status, diff, and uninstall easy. This keeps the existing target abstractions while moving lifecycle behavior behind ordered registries and explicit command registration.
