+++ src/ai_rules/cli/__main__.py
@@ -1,0 +1,8 @@
+"""Module execution entry point for `python -m ai_rules.cli`."""
+
+from __future__ import annotations
+
+from ai_rules.cli import main
+
+if __name__ == "__main__":
+    main()
