+++ src/ai_rules/cli/commands/__init__.py
@@ -1,0 +1,1 @@
+"""Top-level CLI commands."""
