+++ src/ai_rules/cli/groups/__init__.py
@@ -1,0 +1,1 @@
+"""CLI command groups."""
